requirements.txt
fire
intersight
python-dotenv
requests
urllib3
